dev_requirements.txt
SQLAlchemy==0.9.7
PyYAML==3.11
coverage==3.7.1
fudge==1.0.3
mock==1.0.1
nose==1.3.3
